add a dependency on deps_target_dependency that is itself a CMake imported target
The existing dependency was added manually, but discovered through
pkg-config. This dependency only uses the CMake export/import mechanism.

Internally (to CMake), it gets listed in the library's interface link
libraries as-is, i.e. using its target name. This allows us to ensure
that the base/cmake code handles it properly when exported to CMake
(i.e. resolves the library and puts its full path)